:lipstick: split description form
